+++ IoTPublishShare/src/com/aalto/protocol/design/iotps/objects/IoTPSAckObject.java
@@ -2,22 +2,22 @@ package com.aalto.protocol.design.iotps.objects;
 
 public class IoTPSAckObject {
 
-	private double sub_seq_no;
-	private double seq_no;
+	private long sub_seq_no;
+	private long seq_no;
 	
 	private String fromIp;
 	private int fromPort;
 	
-	public double getSubSeqNo() {
+	public long getSubSeqNo() {
 		return sub_seq_no;
 	}
-	public void setSubSeqNo(double sub_seq_no) {
+	public void setSubSeqNo(long sub_seq_no) {
 		this.sub_seq_no = sub_seq_no;
 	}
-	public double getSeqNo() {
+	public long getSeqNo() {
 		return seq_no;
 	}
-	public void setSeqNo(double seq_no) {
+	public void setSeqNo(long seq_no) {
 		this.seq_no = seq_no;
 	}
 	public String getFromIp() {
