+++ IoTPublishShare/src/com/aalto/protocol/design/iotps/objects/IoTPSClientObject.java
@@ -4,8 +4,8 @@ public class IoTPSClientObject extends IoTPSObject {
 
 	private String deviceId;
 	private int port;
-	private double sub_seq_no;
-	private double seq_no;
+	private long sub_seq_no;
+	private long seq_no;
 	private int version;
 	private int ack_support;
 	private String ip;
@@ -22,16 +22,16 @@ public class IoTPSClientObject extends IoTPSObject {
 	public void setPort(int port) {
 		this.port = port;
 	}
-	public double getSubSeqNo() {
+	public long getSubSeqNo() {
 		return sub_seq_no;
 	}
-	public void setSubSeqNo(double sub_seq_no) {
+	public void setSubSeqNo(long sub_seq_no) {
 		this.sub_seq_no = sub_seq_no;
 	}
-	public double getSeqNo() {
+	public long getSeqNo() {
 		return seq_no;
 	}
-	public void setSeqNo(double seq_no) {
+	public void setSeqNo(long seq_no) {
 		this.seq_no = seq_no;
 	}
 	public int getVersion() {
