+++ IoTPublishShare/src/com/aalto/protocol/design/iotps/objects/IoTPSPendingAcksObject.java
@@ -2,19 +2,19 @@ package com.aalto.protocol.design.iotps.objects;
 
 class IoTPSPendingAcksObject extends IoTPSObject {
 
-	private double sub_seq_no;
-	private double seq_no;
+	private long sub_seq_no;
+	private long seq_no;
 	
-	public double getSubSeqNo() {
+	public long getSubSeqNo() {
 		return sub_seq_no;
 	}
-	public void setSubSeqNo(double sub_seq_no) {
+	public void setSubSeqNo(long sub_seq_no) {
 		this.sub_seq_no = sub_seq_no;
 	}
-	public double getSeqNo() {
+	public long getSeqNo() {
 		return seq_no;
 	}
-	public void setSeqNo(double seq_no) {
+	public void setSeqNo(long seq_no) {
 		this.seq_no = seq_no;
 	}
 }
